+++ source/main/modules/fsal/util/sort-natural.ts
@@ -16,7 +16,7 @@ import {
   CodeFileDescriptor,
   DirDescriptor,
   MDFileDescriptor
-} from '../types'
+} from '@dts/main/fsal'
 
 type FileNameDisplay = 'filename'|'title'|'heading'|'title+heading'