+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Tests now simply transpile without type checking
 - Move service provider types to the new types directory
 - Provide `@dts` as an alias to retrieve typings
+- Move FSAL types to the new types directory
 
 # 2.1.2